Presentation on the interchain at EthBerlin 2019
Abstract:

**The interchain cometh! The radical implications of having *every* blockchain in the #defi and dapp ecosystem.**

The interchain, an anticipated "internet of blockchains," promises to realise untapped potential across currently-isolated chains, to resolve some scaling and fit-to-purpose issues, and to usher in a new era of the internet. However, even though some prominent blockchain interoperability projects have launched, some of the fundamental concepts of the interchain (like what constitutes its "edge") are barely explored, let alone well-established. This has created several adverse knock-on effects, like (well-founded) concerns as to whether interoperability projects would interfere with chains' sovereignty or finality, or whether they are in fact Ethereum competitors instead of systemic enablers.

This talk introduces the fundamentals of the interchain, and clarifies how "interoperability" strategies are not usually purely "interchain," but tend to combine regular blockchain tech with it. I suggest how interchain projects can safely avoid pitfalls like competing with the chains they hope to work with. Finally, I make the case that the interchain presents our single biggest opportunity to enable massive systemic ecosystem benefits - but this is premised on our avoiding a tragedy of the commons where the adoption of competitive or parasitic interoperability projects creates a mere turf war, instead of fostering a true, unopinionated interchain. Imagine TCP/IP were to privilege one website over another; we need to actively avoid the equivalent in the blockchain space.